% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{coef.occu_inla}
\alias{coef.occu_inla}
\title{Extract model coefficients}
\usage{
\method{coef}{occu_inla}(object, process = c("both", "occupancy", "detection"), ...)
}
\arguments{
\item{object}{fitted occu_inla object}

\item{process}{\code{"occupancy"} (default), \code{"detection"}, or
\code{"both"}.}

\item{...}{ignored}
}
\value{
Named numeric vector of posterior means. If
\code{process = "both"}, a named list with \code{occ} and \code{det}.
}
\description{
Extract model coefficients
}
