% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{compare_models}
\alias{compare_models}
\title{Compare two occupancy models via WAIC}
\usage{
compare_models(...)
}
\arguments{
\item{...}{named occu_inla objects to compare}
}
\value{
data.frame with model comparison metrics
}
\description{
Compare two occupancy models via WAIC
}
