% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_diagnostics.R
\name{fitted.occu_inla_jsdm}
\alias{fitted.occu_inla_jsdm}
\title{Extract fitted values from a JSDM}
\usage{
\method{fitted}{occu_inla_jsdm}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_jsdm object}

\item{...}{ignored}
}
\value{
named list of per-species fitted probability vectors
}
\description{
Extract fitted values from a JSDM
}
