% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_diagnostics.R
\name{fitted.occu_inla_ms}
\alias{fitted.occu_inla_ms}
\title{Extract fitted values from a multi-species occupancy model}
\usage{
\method{fitted}{occu_inla_ms}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_ms object}

\item{...}{ignored}
}
\value{
named list of per-species fitted value lists
}
\description{
Extract fitted values from a multi-species occupancy model
}
