% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_diagnostics.R
\name{fitted.occu_inla_temporal}
\alias{fitted.occu_inla_temporal}
\title{Extract fitted values from a temporal occupancy model}
\usage{
\method{fitted}{occu_inla_temporal}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_temporal object}

\item{...}{ignored}
}
\value{
list of per-period fitted value lists
}
\description{
Extract fitted values from a temporal occupancy model
}
