% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_diagnostics.R
\name{getSVCSamples}
\alias{getSVCSamples}
\title{Extract spatially-varying coefficient summaries}
\usage{
getSVCSamples(object)
}
\arguments{
\item{object}{fitted occu_inla_svc object}
}
\value{
list of data.frames (one per SVC) with site-level mean, sd, quantiles
}
\description{
Extract spatially-varying coefficient summaries
}
