% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAocc-package.R
\name{glance}
\alias{glance}
\title{Glance at model-level statistics}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{a model object}

\item{...}{additional arguments passed to methods}
}
\value{
A one-row data.frame.
}
\description{
Generic for producing single-row model summaries. See
\code{\link{glance.occu_inla}} for the occupancy model method.
}
