% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{logLik.occu_inla}
\alias{logLik.occu_inla}
\alias{logLik.occu_inla_temporal}
\title{Extract log-likelihood}
\usage{
\method{logLik}{occu_inla}(object, ...)

\method{logLik}{occu_inla_temporal}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla object}

\item{...}{ignored}
}
\value{
An object of class \code{"logLik"}.
}
\description{
Returns the observed-data log-likelihood from the final EM iteration.
The \code{df} attribute is set to the number of fixed effect
coefficients (occupancy + detection) and the \code{nobs} attribute
to the number of non-NA detection history entries.
}
