% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{nobs.occu_inla}
\alias{nobs.occu_inla}
\title{Number of observations}
\usage{
\method{nobs}{occu_inla}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla object}

\item{...}{ignored}
}
\value{
Integer: number of non-NA detection history entries.
}
\description{
Number of observations
}
