% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_data.R
\name{occu_format_ms}
\alias{occu_format_ms}
\title{Create multi-species occupancy data}
\usage{
occu_format_ms(y_list, occ.covs = NULL, det.covs = NULL, coords = NULL)
}
\arguments{
\item{y_list}{named list of N x J detection matrices (one per species)}

\item{occ.covs}{site-level covariates (shared across species)}

\item{det.covs}{detection covariates (shared or species-specific)}

\item{coords}{optional coordinates}
}
\value{
object of class \code{"occu_data_ms"}
}
\description{
Create multi-species occupancy data
}
