% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_utils.R
\name{occu_priors}
\alias{occu_priors}
\title{Specify priors for INLA occupancy models (spOccupancy-compatible)}
\usage{
occu_priors(
  beta.normal = list(mean = 0, var = 2.72),
  alpha.normal = list(mean = 0, var = 2.72),
  sigma.sq.psi.ig = c(0.1, 0.1),
  sigma.sq.p.ig = c(0.1, 0.1),
  sigma.sq.ig = NULL,
  phi.unif = NULL,
  beta.comm.normal = NULL,
  alpha.comm.normal = NULL,
  tau.sq.beta.ig = NULL,
  tau.sq.alpha.ig = NULL
)
}
\arguments{
\item{beta.normal}{list(mean, var) for occupancy fixed effects Normal prior}

\item{alpha.normal}{list(mean, var) for detection fixed effects Normal prior}

\item{sigma.sq.psi.ig}{c(shape, scale) for occupancy RE variance Inv-Gamma}

\item{sigma.sq.p.ig}{c(shape, scale) for detection RE variance Inv-Gamma}

\item{sigma.sq.ig}{c(shape, scale) for spatial variance (spatial models)}

\item{phi.unif}{c(lower, upper) for spatial decay Uniform prior}

\item{beta.comm.normal}{list(mean, var) community occ prior (multi-species)}

\item{alpha.comm.normal}{list(mean, var) community det prior (multi-species)}

\item{tau.sq.beta.ig}{list(a, b) species-level occ variance (multi-species)}

\item{tau.sq.alpha.ig}{list(a, b) species-level det variance (multi-species)}
}
\value{
list of class \code{"occu_priors"} with INLA-compatible prior specs
}
\description{
Specify priors for INLA occupancy models (spOccupancy-compatible)
}
