% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{plot.occu_inla}
\alias{plot.occu_inla}
\title{Plot diagnostics for occu_inla}
\usage{
\method{plot}{occu_inla}(x, which = 1:4, ...)
}
\arguments{
\item{x}{fitted occu_inla object}

\item{which}{integer vector: which plots to show.
1 = EM convergence, 2 = psi histogram, 3 = p histogram,
4 = psi vs covariates.}

\item{...}{additional arguments passed to plot}
}
\value{
The \code{occu_inla} object \code{x}, returned invisibly.
}
\description{
Plot diagnostics for occu_inla
}
