% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{plot.occu_inla_jsdm}
\alias{plot.occu_inla_jsdm}
\title{Plot diagnostics for JSDM}
\usage{
\method{plot}{occu_inla_jsdm}(x, which = 1:2, ...)
}
\arguments{
\item{x}{fitted occu_inla_jsdm object}

\item{which}{integer vector: which plots to show.
1 = species occupancy, 2 = factor loadings heatmap.}

\item{...}{additional arguments passed to plot}
}
\value{
The object \code{x}, returned invisibly.
}
\description{
Plot diagnostics for JSDM
}
