% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{plot.occu_inla_ms}
\alias{plot.occu_inla_ms}
\title{Plot diagnostics for multi-species occupancy model}
\usage{
\method{plot}{occu_inla_ms}(x, which = 1:2, ...)
}
\arguments{
\item{x}{fitted occu_inla_ms object}

\item{which}{integer vector: which plots to show.
1 = per-species occupancy, 2 = per-species detection,
3 = community effects.}

\item{...}{additional arguments passed to plot}
}
\value{
The object \code{x}, returned invisibly.
}
\description{
Plot diagnostics for multi-species occupancy model
}
