% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{plot.occu_inla_temporal}
\alias{plot.occu_inla_temporal}
\title{Plot diagnostics for temporal occupancy model}
\usage{
\method{plot}{occu_inla_temporal}(x, which = 1:2, ...)
}
\arguments{
\item{x}{fitted occu_inla_temporal object}

\item{which}{integer vector: which plots to show.
1 = occupancy over time, 2 = detection over time.}

\item{...}{additional arguments passed to plot}
}
\value{
The object \code{x}, returned invisibly.
}
\description{
Plot diagnostics for temporal occupancy model
}
