% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_predict.R
\name{plot.occu_prediction}
\alias{plot.occu_prediction}
\title{Plot predicted effects from an occupancy model}
\usage{
\method{plot}{occu_prediction}(x, ...)
}
\arguments{
\item{x}{an \code{occu_prediction} object from
\code{predict(fit, terms = ...)}}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}}
}
\value{
The input object, invisibly. Called for side effect of producing
a plot.
}
\description{
Plot predicted effects from an occupancy model
}
