% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_predict.R
\name{predict.occu_inla_jsdm}
\alias{predict.occu_inla_jsdm}
\title{Predict from a JSDM}
\usage{
\method{predict}{occu_inla_jsdm}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_jsdm object}

\item{...}{additional arguments (ignored)}
}
\value{
named list of per-species occupancy probability vectors
}
\description{
Returns per-species occupancy predictions (no detection process).
}
