% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_predict.R
\name{predict_spatial}
\alias{predict_spatial}
\title{Predict occupancy at new spatial locations}
\usage{
predict_spatial(
  object,
  newcoords,
  newocc.covs = NULL,
  ignore.RE = FALSE,
  n.samples = 500
)
}
\arguments{
\item{object}{fitted occu_inla_spatial object}

\item{newcoords}{n_pred x 2 matrix of prediction coordinates}

\item{newocc.covs}{data.frame of occupancy covariates at prediction locations}

\item{ignore.RE}{ignore random effects?}

\item{n.samples}{number of posterior samples}
}
\value{
list with predicted psi at new locations
}
\description{
Convenience wrapper for spatial models. Constructs X.0 from a covariate
data.frame and handles coordinate projection.
}
