% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_data.R
\name{print.occu_data}
\alias{print.occu_data}
\title{Print method for occu_data}
\usage{
\method{print}{occu_data}(x, ...)
}
\arguments{
\item{x}{an \code{occu_data} object}

\item{...}{additional arguments (ignored)}
}
\value{
The \code{occu_data} object \code{x}, returned invisibly.
}
\description{
Print method for occu_data
}
