% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{print.occu_inla}
\alias{print.occu_inla}
\title{Print method for occu_inla}
\usage{
\method{print}{occu_inla}(x, ...)
}
\arguments{
\item{x}{fitted occu_inla object}

\item{...}{additional arguments (ignored)}
}
\value{
The \code{occu_inla} object \code{x}, returned invisibly.
}
\description{
Print method for occu_inla
}
