% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{print.occu_inla_int}
\alias{print.occu_inla_int}
\title{Print method for integrated occupancy model}
\usage{
\method{print}{occu_inla_int}(x, ...)
}
\arguments{
\item{x}{fitted occu_inla_int object}

\item{...}{additional arguments (ignored)}
}
\value{
The object \code{x}, returned invisibly.
}
\description{
Print method for integrated occupancy model
}
