% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{print.occu_inla_jsdm}
\alias{print.occu_inla_jsdm}
\title{Print method for JSDM}
\usage{
\method{print}{occu_inla_jsdm}(x, ...)
}
\arguments{
\item{x}{fitted occu_inla_jsdm object}

\item{...}{additional arguments (ignored)}
}
\value{
The object \code{x}, returned invisibly.
}
\description{
Print method for JSDM
}
