% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{print.occu_inla_ms}
\alias{print.occu_inla_ms}
\title{Print method for multi-species occupancy model}
\usage{
\method{print}{occu_inla_ms}(x, ...)
}
\arguments{
\item{x}{fitted occu_inla_ms object}

\item{...}{additional arguments (ignored)}
}
\value{
The object \code{x}, returned invisibly.
}
\description{
Print method for multi-species occupancy model
}
