% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_spatial.R
\name{print.occu_spatial}
\alias{print.occu_spatial}
\title{Print method for occu_spatial}
\usage{
\method{print}{occu_spatial}(x, ...)
}
\arguments{
\item{x}{an \code{occu_spatial} object}

\item{...}{additional arguments (ignored)}
}
\value{
The \code{occu_spatial} object \code{x}, returned invisibly.
}
\description{
Print method for occu_spatial
}
