% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAocc-package.R
\name{ranef}
\alias{ranef}
\title{Extract random effects}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{additional arguments passed to methods}
}
\value{
A list of data.frames.
}
\description{
Generic for extracting random effect estimates. See
\code{\link{ranef.occu_inla}} for the occupancy model method.
}
