% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_predict.R
\name{richness}
\alias{richness}
\title{Compute species richness from multi-species model}
\usage{
richness(object)
}
\arguments{
\item{object}{fitted occu_inla_ms object}
}
\value{
data.frame with site-level richness estimates and uncertainty
}
\description{
Compute species richness from multi-species model
}
