% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_data.R
\name{simMsOcc}
\alias{simMsOcc}
\title{Simulate multi-species occupancy data (cf. simMsOcc)}
\usage{
simMsOcc(
  N = 100,
  J = 4,
  n_species = 10,
  n_occ_covs = 1,
  n_det_covs = 1,
  beta_comm_mean = NULL,
  beta_comm_sd = NULL,
  alpha_comm_mean = NULL,
  alpha_comm_sd = NULL,
  spatial_range = NULL,
  spatial_var = 1,
  seed = NULL
)
}
\arguments{
\item{N}{number of sites}

\item{J}{number of visits}

\item{n_species}{number of species}

\item{n_occ_covs}{number of occupancy covariates}

\item{n_det_covs}{number of detection covariates}

\item{beta_comm_mean}{community mean for occupancy coefficients}

\item{beta_comm_sd}{community SD for occupancy coefficients}

\item{alpha_comm_mean}{community mean for detection coefficients}

\item{alpha_comm_sd}{community SD for detection coefficients}

\item{spatial_range}{spatial range (NULL = no spatial effect)}

\item{spatial_var}{spatial variance}

\item{seed}{random seed}
}
\value{
list with occu_data_ms object and true parameters
}
\description{
Simulate multi-species occupancy data (cf. simMsOcc)
}
