% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_data.R
\name{simTMsOcc}
\alias{simTMsOcc}
\title{Simulate temporal multi-species occupancy data (cf. simTMsOcc)}
\usage{
simTMsOcc(
  N = 100,
  J = 4,
  n_species = 10,
  n_seasons = 5,
  n_occ_covs = 1,
  n_det_covs = 1,
  beta_comm_mean = NULL,
  beta_comm_sd = NULL,
  alpha_comm_mean = NULL,
  alpha_comm_sd = NULL,
  ar1 = TRUE,
  rho = 0.7,
  sigma_t = 0.5,
  seed = NULL
)
}
\arguments{
\item{N}{number of sites}

\item{J}{number of visits per season}

\item{n_species}{number of species}

\item{n_seasons}{number of primary periods}

\item{n_occ_covs}{number of occupancy covariates}

\item{n_det_covs}{number of detection covariates}

\item{beta_comm_mean}{community mean for occupancy coefficients}

\item{beta_comm_sd}{community SD for occupancy coefficients}

\item{alpha_comm_mean}{community mean for detection coefficients}

\item{alpha_comm_sd}{community SD for detection coefficients}

\item{ar1}{logical: use AR(1) temporal correlation}

\item{rho}{AR(1) correlation parameter}

\item{sigma_t}{temporal innovation SD}

\item{seed}{random seed}
}
\value{
list with 4D data and true parameters
}
\description{
Simulate temporal multi-species occupancy data (cf. simTMsOcc)
}
