% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{summary.occu_inla_int}
\alias{summary.occu_inla_int}
\title{Summary for integrated occupancy model}
\usage{
\method{summary}{occu_inla_int}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_int object}

\item{...}{additional arguments (ignored)}
}
\value{
Invisibly returns a summary list.
}
\description{
Summary for integrated occupancy model
}
