% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{summary.occu_inla_jsdm}
\alias{summary.occu_inla_jsdm}
\title{Summary for JSDM}
\usage{
\method{summary}{occu_inla_jsdm}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_jsdm object}

\item{...}{additional arguments (ignored)}
}
\value{
Invisibly returns a per-species summary data.frame.
}
\description{
Summary for JSDM
}
