% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_output.R
\name{summary.occu_inla_ms}
\alias{summary.occu_inla_ms}
\title{Summary for multi-species model}
\usage{
\method{summary}{occu_inla_ms}(object, ...)
}
\arguments{
\item{object}{fitted occu_inla_ms object}

\item{...}{additional arguments (ignored)}
}
\value{
The multi-species model object, returned invisibly.
}
\description{
Summary for multi-species model
}
