% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAocc-package.R
\name{tidy}
\alias{tidy}
\title{Tidy model output into a data.frame}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{a model object}

\item{...}{additional arguments passed to methods}
}
\value{
A data.frame.
}
\description{
Generic for producing tidy coefficient tables. See
\code{\link{tidy.occu_inla}} for the occupancy model method.
}
